requirements_web.txt
aioredis
appdirs
asgiref
async-timeout
atomicwrites
attrs
audioread
autobahn
Automat
bcrypt
certifi
cffi
channels
channels-redis
chardet
colorama
constantly
cryptography
cycler
daphne
decorator
Django
django-better-admin-arrayfield
django-debug-toolbar
django-cors-headers
djangorestframework
filelock
hiredis
huggingface-hub
hyperlink
HyperPyYAML
idna
incremental
iniconfig
joblib
kiwisolver
librosa
llvmlite
matplotlib
msgpack
nanoid
numba
numpy
packaging
pandas
Pillow
pip
pluggy
pooch
psycopg2
py
pyasn1
pyasn1-modules
pycparser
pyOpenSSL
pyparsing
pytest
python-dateutil
pytz
PyYAML
requests
resampy
ruamel.yaml.clib
ruamel.yaml
scikit-learn
scipy
sentencepiece
service-identity
setuptools
six
SoundFile
spafe
speechbrain
sqlparse
threadpoolctl
toml
torch
torchaudio
torchvision
tqdm
Twisted
txaio
typing-extensions
urllib3
zope.interface
natasha==0.10.0
pymorphy2==0.8
yargy==0.12.0